backend/requirements.txt
# flask
# flask-cors
# logging
# werkzeug
# torch 
# transformers 
# indic-nlp-library
# sentencepiece 
# sacremoses
# numpy
# pytesseract
# easyocr
# os
# uuid
# logging
# pandas
# indicnlp
# transformers
# pillow
# Image
Flask==2.3.2
flask-cors==4.0.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
Pillow==10.3.0
pytesseract==0.3.10
torch==2.2.2
easyocr==1.7.1
transformers==4.41.1
indic-nlp-library==0.92
sentencepiece==0.2.0
tokenizers==0.19.1
protobuf==4.25.3